+++ bin/umap_reduce.py
@@ -5,6 +5,7 @@ import logging
 from pathlib import Path
 
 import pandas as pd
+from sklearn.decomposition import IncrementalPCA
 import umap
 
 
@@ -18,8 +19,8 @@ def parse_args():
     # Positional mandatory arguments
     parser.add_argument(
         "matrix",
-        help="Gene expression matrix: rows=genes, columns=barcodes, \
-        values=UMIs",
+        help="Gene expression matrix: rows=genes/transcripts, \
+            columns=barcodes, values=UMIs",
     )
 
     # Optional arguments
@@ -27,8 +28,16 @@ def parse_args():
         "-o",
         "--output_prefix",
         help="write to current directory with this prefix",
-        type=str,
         default="")
+
+    parser.add_argument(
+        "-p",
+        "--pcn",
+        help="Number of principal components to generate prior to UMAP \
+        dimensionality reduction [100]",
+        type=int,
+        default=100,
+    )
 
     parser.add_argument(
         "-d",
@@ -36,13 +45,6 @@ def parse_args():
         help="Number of dimensions to reduce to [2]",
         type=int,
         default=2,
-    )
-
-    parser.add_argument(
-        "-f",
-        "--feature_type",
-        help="matrix is gene or transcript [gene]",
-        default='gene',
     )
 
     parser.add_argument(
@@ -91,17 +93,36 @@ def init_logger(args):
     logging.root.handlers[0].addFilter(lambda x: "NumExpr" not in x.msg)
 
 
+def pca(x, args):
+    """Run PCA to generate args.pcn principal components."""
+    pcn = min(args.pcn, x.shape[1])  # pcn must not exceed N features
+    transformer = IncrementalPCA(n_components=pcn)
+    return transformer.fit_transform(x)
+
+
 def main(args):
     """Run entry point."""
     init_logger(args)
 
-    df = pd.read_csv(args.matrix, delimiter="\t").set_index(args.feature_type)
+    df = pd.read_csv(args.matrix, delimiter="\t")
+    feature_header = df.columns[0]
+    if feature_header == "gene":
+        df = df.set_index("gene")
+    elif feature_header == "transcript":
+        df = df.set_index("transcript")
+    else:
+        raise Exception("Did not find expected gene/transcript header")
 
     # Switch from barcodes as columns to barcodes as rows
     x = df.transpose()
 
     logger.info(
-        f"Running UMAP: {x.shape[1]} features --> \
+        f"Running PCA: {x.shape[1]} features --> \
+            {args.pcn} prinicipal components")
+    x_pca = pca(x, args)
+
+    logger.info(
+        f"Running UMAP: {x_pca.shape[1]} features --> \
             {args.dimensions} dimensions")
 
     for n in range(args.num_umaps):
@@ -113,18 +134,17 @@ def main(args):
         )
         outpath = Path() / f"{args.output_prefix}_{n}_umap.tsv"
 
-        # For testing: If there's only a single transcript column the reducer
+        # If there's only a single transcript column, the reducer
         # step will fail. For now just write an empty file.
         # TODO: Have a better way of detcting if transcript data does not have
         # enough transciript columns. Probably only an issue with test data
         try:
-            x_embedded = reducer.fit_transform(x)
+            x_embedded = reducer.fit_transform(x_pca)
         except TypeError:
             open(outpath, 'w').close()
         else:
             cols = [f"D{i+1}" for i in range(args.dimensions)]
-            df_umap = pd.DataFrame(
-                x_embedded, columns=cols, index=x.index)
+            df_umap = pd.DataFrame(x_embedded, columns=cols, index=x.index)
 
             df_umap.to_csv(
                 outpath, sep="\t", index=True, index_label="barcode")